% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addItem.R
\name{addItem}
\alias{addItem}
\title{Add an Item to an Existing Closet}
\usage{
addItem(
  brand,
  color,
  description,
  type = c("top", "bottom", "one-piece", "shoes"),
  file.name
)
}
\arguments{
\item{brand}{item brand.}

\item{color}{item color.}

\item{description}{description of the item.}

\item{type}{must be one of "top", "bottom", "one-piece" (such as a dress or jumpsuit), or
"shoes".}

\item{file.name}{name(s) of closet files to add item to.}
}
\value{

}
\description{
Add an item to one or more closet files.
}
\examples{
addItem(brand = "Gap", color = "green", description = "sweater",
  type = "top", file.name = "work_closet.csv")
    # the item will print elsewhere as: brand color description. So this
    # item prints as 'Gap green sweater'.

addItem(brand = "J. Crew", color = "black",
  description = "sleeveless jumpsuit", type = "one-piece",
  file.name = c("work_closet.csv","fancy_closet.csv"))
}
\seealso{
\code{\link{newCloset}}, \code{\link{loadCloset}},
  \code{\link{randomOutfit}}
}
