% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadCloset.R
\name{loadCloset}
\alias{loadCloset}
\title{Load an Existing Closet}
\usage{
loadCloset(file.name)
}
\arguments{
\item{file.name}{name of the closet csv file to load.}
}
\value{

}
\description{
Load a closet dataframe from a csv file. The file must include four variables
with the headers: 'brand', 'color', 'description', and 'type' (one of 'top',
'bottom', 'one-piece', or 'shoes'). It can be generated using
\code{\link{newCloset}} and \code{\link{addItem}}.
}
\examples{
work <- loadCloset("work_closet.csv")
}
\seealso{
\code{\link{newCloset}}, \code{\link{addItem}},
  \code{\link{randomOutfit}}
}
