% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newCloset.R
\name{newCloset}
\alias{newCloset}
\title{Create a New Closet}
\usage{
newCloset(file.name)
}
\arguments{
\item{file.name}{unique file path/name(s) for new csv file(s) with closet
variable names.}
}
\description{
Creates one or more new csv files with the headers 'brand', 'color',
'description', and 'type'.
}
\examples{
newCloset(file.name = "work_closet.csv")

newCloset(file.name = c("fancy_closet.csv","casual_closet.csv"))
}
\seealso{
\code{\link{addItem}}, \code{\link{loadCloset}},
  \code{\link{randomOutfit}}
}
