% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pickOutfit.R
\name{pickOutfit}
\alias{pickOutfit}
\title{Pick a Random Outfit}
\usage{
pickOutfit(closet.file, style = NULL)
}
\arguments{
\item{closet.file}{the name of the file containing your closet inventory. See 
\code{\link{addItem}} to build a closet.}

\item{style}{(optional) category from which to select items. The default 
\code{style = NULL} picks an outfit from all items in your closet.}
}
\value{
Returns a character string outfit consisting of a top and a bottom, or a 
    one-piece outfit (such as a dress or jumpsuit), and a pair of shoes.
}
\description{
Pick a Random Outfit
}
\examples{
pickOutfit("closet.txt", style = "work")
}
\seealso{
\code{\link{addItem}}
}
