% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomOutfit.R
\name{randomOutfit}
\alias{randomOutfit}
\title{Pick a Random Outfit}
\usage{
randomOutfit(closet.name)
}
\arguments{
\item{closet.name}{a closet dataframe containing four variables: 'brand',
'color', 'description', and 'type'.}
}
\value{
An outfit.
}
\description{
Randomly select a top and a bottom or a one-piece outfit (such as a dress or
jumpsuit) and a pair of shoes from the closet dataframe.
}
\examples{
work <- loadCloset("work_closet.csv")
randomOutfit(work)

# get an outfit without storing the closet in the R environment
randomOutfit(loadCloset("fancy_closet.csv"))
}
\seealso{
\code{\link{newCloset}}, \code{\link{addItem}},
  \code{\link{loadCloset}}
}
