% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addItem.R
\name{addItem}
\alias{addItem}
\title{Add an Item to Your Closet}
\usage{
addItem(
  item,
  type = c("top", "bottom", "one-piece", "shoes"),
  style = NA,
  closet.file
)
}
\arguments{
\item{item}{description of the item.}

\item{type}{must be one of \code{"top"}, \code{"bottom"}, \code{"one-piece"}
(such as a dress or jumpsuit), or \code{"shoes"}.}

\item{style}{(optional) one or more style category tags to apply to the item.
Items within a style can be mixed-and-matched (for example, a
capsule wardrobe). Items without an assigned style go with everything!}

\item{closet.file}{the name of the file containing your closet inventory. If
the file does not exist, it will be created.}
}
\value{

}
\description{
Add an Item to Your Closet
}
\examples{
addItem(item = "green skirt", type = "bottom", style = "work",
    closet.file = "closet.txt")

addItem(item = "black sleeveless jumpsuit", type = "one-piece",
    style = c("work","fancy"), closet.file = "closet.txt")

addItem(item = "black t-shirt", type = "top",
    closet.file = "closet.txt")

addItem(item = "brown suede ankle boots", type = "shoes",
    closet.file = "closet.txt")
}
\seealso{
\code{\link{pickOutfit}}
}
